+++ changelog.md
@@ -1,6 +1,8 @@
 # Version 0.2
 - FEATURE: Added '/cgive pastebin [id]' command
 - FEATURE: Added '/cobalt reload' command
+- FEATURE: Added '/edit item model [model]' command
+- FEATURE: Added '/edit item name [name]' command
 
 # Version 0.1
 Initial version